컨트롤러가 단일 배치 서비스만 보도록 정리한다
배치 실행과 상태 이력 조회 로직을 BatchManualRunService로 모아 BatchApiController가 배치 서비스 하나만 주입받도록 정리했다.

같은 흐름에서 ISO_LOCAL_DATE_TIME 포매터도 공용 Constants로 옮겨 표현 상수를 한곳에 모았다.

Constraint: 컨트롤러는 배치 관련 유스케이스를 하나의 서비스 경계로 바라보는 편이 단순했다
Constraint: 상태 조회 기능을 추가하면서 컨트롤러에 서비스 의존이 불필요하게 늘어나지 않아야 했다
Rejected: BatchStatusQueryService를 별도 주입으로 유지 | 컨트롤러가 배치 유스케이스 두 갈래를 직접 조합하게 됨
Rejected: DateTimeFormatter를 컨트롤러 지역 상수로 유지 | 표현 포맷 상수가 흩어져 재사용성이 떨어짐
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: 배치 관련 읽기/쓰기 유스케이스를 더 늘릴 때도 컨트롤러 의존을 먼저 늘리기보다 기존 배치 서비스 경계 안에서 통합 가능한지 검토할 것
Tested: ./gradlew :apps:batch-core:test :apps:dragons_api:test --tests com.dragons.interfaces.api.ApiControllerTest
Not-tested: 배치 서비스가 더 많은 조회 책임을 가질 때의 장기적인 분리 기준 재검토
